Refactor persons and rentals management part
